requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
dataclasses-json==0.6.3
defusedxml==0.8.0rc2
Django==5.0.1
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.9
langsmith==0.0.79
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nanoid==2.0.0
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pgvector==0.2.4
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
